backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-environ
celery[redis]
django-cors-headers
requests
django-redis
